design : edit PostHashTag position #46
해시태그 추가/삭제 부분 위치를 수정합니다.
해시태그 추가/삭제 부분을 컴포넌트화 합니다.
