+++ src/components/PostTitle.tsx
@@ -18,16 +18,6 @@ const regionOptions = [
   { value: "광주", label: "광주" },
   { value: "울산", label: "울산" },
 ];
-const hashTagOptions = [
-  { value: "#여자혼자", label: "#여자혼자" },
-  { value: "#남자혼자", label: "#남자혼자" },
-  { value: "#커플끼리", label: "#커플끼리" },
-  { value: "#부부끼리", label: "#부부끼리" },
-  { value: "#부모님과", label: "#부모님과" },
-  { value: "#반려동물", label: "#반려동물" },
-  { value: "#아이들과", label: "#아이들과" },
-  { value: "#단체여행", label: "#단체여행" },
-];
 
 const PostTitle = () => {
   return (
@@ -41,15 +31,6 @@ const PostTitle = () => {
           />
         </div>
         <input className="w-3/4 md:w-6/7 md:text-3xl sm:text-lg sm:w-full p-2 border-slate-300 border-b-2 " />
-      </div>
-      <div>
-        <Select
-          options={hashTagOptions}
-          isMulti
-          isSearchable={false}
-          isClearable={true}
-          placeholder={"#해시태그"}
-        />
       </div>
     </div>
   );
