+++ src/pages/Post.tsx
@@ -1,10 +1,12 @@
 import PostBtn from "../components/PostBtn";
+import PostHashTag from "../components/PostHashTag";
 import PostTitle from "../components/PostTitle";
 
 const Post = () => {
   return (
     <>
       <PostTitle />
+      <PostHashTag />
       <PostBtn />
     </>
   );
